+++ libs/base_core/src/text/layout/mod.rs
@@ -1,0 +1,22 @@
+// devela_base_core::text::layout
+//
+#![doc = crate::_DOC_TEXT_LAYOUT!()]
+//!
+//
+
+#[cfg(test)]
+mod tests;
+
+mod engine; // TextFit, TextLayout, TextLayoutStep
+mod symbol; // TextCohesion, TextCursor, TextSpan, TextSymbol,
+mod unit; // TextUnit, TextIndex
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            engine::*,
+            symbol::*,
+            unit::*,
+        };
+    }
+}