+++ libs/base_text/src/text/mod.rs
@@ -5,12 +5,17 @@
 
 #[allow(hidden_glob_reexports, reason = "re-exported `char`")]
 mod char; // transliterate
+
 pub mod grapheme; // Grapheme[Nonul|U8], Grapheme[Boundary|Machine|Prop[Cb|InCb|s]|Scanner]
 
-crate::structural_mods! { // _mods, _workspace_internals
+crate::structural_mods! { // _mods, _pub_mods, _workspace_internals
     _mods {
         pub use super::{
             char::_all::*,
+        };
+    }
+    _pub_mods {
+        pub use super::{
             grapheme::_all::*,
         };
     }