+++ src/text/layout/mod.rs
@@ -1,0 +1,11 @@
+// devela::text::layout
+//
+#![doc = crate::_DOC_TEXT_LAYOUT!()]
+//!
+//
+
+crate::structural_mods! { // _reexports
+    _reexports {
+        pub use devela_base_core::text::layout::*;
+    }
+}